feauture: add rank

+++ src/main/kotlin/it/necross/interfaces/rank/IBoardLastMatch.kt
@@ -1,3 +1,3 @@
 package it.necross.interfaces.rank
 
-data class IBoardLastMatch(val result: String, val mmrChange: Int, val skillMeanChange: Int, val skillStdevChange: Int)
+data class IBoardLastMatch(val result: String, val mmrChange: Int, val skillMeanChange: Float, val skillStdevChange: Float)
