+++ src/main/kotlin/it/necross/interfaces/rank/ICurrent.kt
@@ -1,3 +1,3 @@
 package it.necross.interfaces.rank
 
-data class ICurrent(val id: Int, val name: String, val mmr: Int, val icon: String)
+data class ICurrent(val id: Int, val name: String, val mmr: Float, val icon: String)
